// contracts/core/v1/models/UpdateContestRequestModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateContestRequestModel = {
    title?: string;
    description?: string;
    visibility?: string;
    monitor_scope?: string;
    submissions_list_scope?: string;
    submissions_review_scope?: string;
};